Update & Persist Projects
solving this has been a bit of an issue. first had to get project state to load in in somewhat async matter [project] was solution. Then has to groom request body in a way server expects it to come in. Spring JPA only auto updates if you add in the id. Issue with this was we sometimes cross UUID and ID together and refeenrce the uuid as the DB ID so i just had to adjust the request objects body